requirements.txt
streamlit
pandas
numpy
Pillow
openpyxl
gspread
gspread-dataframe
oauth2client
google-api-python-client
